prototypes/team2_energy/requirements.txt
# team2_energy Requirements
torch>=2.0.0
numpy>=1.24.0
pytest>=7.0.0
cupy-cuda12x>=12.0.0
nvidia-ml-py>=12.0.0
